+++ src/app/search/search.component.html
@@ -1,8 +1,5 @@
-<div #mapContainer id="map"></div>
-<div class="container">
-  {{this.calculatedExpenditure | json}}
-
-  <p>radius--{{radius }}</p>
-
+<div class="searchResults">
+  <div #mapContainer id="map"></div>
+  <app-search-list  [applicableDestinations]="this.applicableDestinations"></app-search-list>
 </div>
-<app-search-list  [applicableDestinations]="this.applicableDestinations"></app-search-list>
+
