+++ src/app/search/search.component.ts
@@ -1,12 +1,9 @@
 import {Component, OnInit, AfterViewInit, ViewChild, ElementRef} from '@angular/core';
-import * as constant from '../searchConstants';
 import {MatSnackBar} from '@angular/material';
-import {GlobalDestinationsObject,CalculatedExpenditure,UserParameters} from '../model/search-criteria';
+import {GlobalDestinationsObject,UserParameters} from '../model/search-criteria';
 import LocationData from './location.json';
 import {SearchDataService} from '../services/search-data.serivce';
-import {Subscription} from 'rxjs';
-import {BehaviorSubject} from 'rxjs';
-
+import {BehaviorSubject, Subscription} from 'rxjs';
 
 @Component({
   selector: 'app-search',
@@ -22,12 +19,6 @@ export class SearchComponent implements OnInit, AfterViewInit {
   applicableDestinations: any;
   searchDataSubs: Subscription;
   availableLocationsSubs: Subscription;
-  calculatedExpenditure: CalculatedExpenditure = {
-    hotelExpenditure: 0,
-    foodExpenditure: 0,
-    travelExpenditure: 0
-  };
-  radius: any;
 
   globalDestinationsObject : GlobalDestinationsObject[] = LocationData;
 
@@ -38,8 +29,6 @@ export class SearchComponent implements OnInit, AfterViewInit {
   }
 
   ngOnInit() {
-    // Check why we need to initSearch in afterviewinit
-    //this.userParameters = this.searchDataService.getUserSearchData();
     this.applicableDestinations = new BehaviorSubject(this.applicableLocations);
   }
 
@@ -48,164 +37,27 @@ export class SearchComponent implements OnInit, AfterViewInit {
      * Init search here
      */
     this.searchDataSubs = this.searchDataService.getUserSearchData().subscribe((data) => {
-      this.initSearch(data)
+      if (data) {
+        sessionStorage.setItem('userData', JSON.stringify(data));
+        this.searchDataService.initSearch(data);
+      } else if ( JSON.parse(sessionStorage.getItem('userData')) ) {
+        let userData = JSON.parse(sessionStorage.getItem('userData'));
+        this.searchDataService.initSearch(userData);
+      } else 
+        this.snackBar.open('Please search again!', '', {duration: 5000})
+    });
+
+    // Subscribe to location data
+    this.availableLocationsSubs = this.searchDataService.getApplicableLocationsSubs().subscribe(data => {
+      if(data){
+        console.log('DATA' + data);
+        this.applicableLocations = data.location;
+        this.mapInitializer(this.applicableLocations, data.position);
+        this.applicableDestinations.next(this.applicableLocations);
+        console.log('destinations set in parent');
+      }
     });
   }
-
-  initSearch(userParameters: UserParameters) {
-
-    /**
-     * Initial parameters supplied from User
-     */
-
-    let budget;
-    let byRoad = false;
-
-    if (userParameters.travel.travelmode === constant.travelMode.twowheeler ||
-      userParameters.travel.travelmode === constant.travelMode.bus ||
-      userParameters.travel.travelmode === constant.travelMode.fourwheeler) {
-      byRoad = true;
-    }
-
-    const defaultPosition = {
-      coords: {
-        latitude: constant.searchConstants.defaultLocation.latitude,
-        longitude: constant.searchConstants.defaultLocation.longitude
-      }
-    };
-
-    try {
-
-      budget = userParameters.budget * userParameters.person;
-      this.calculatedExpenditure.hotelExpenditure = this.getHotelExpenses(userParameters, budget);
-      budget = budget - this.calculatedExpenditure.hotelExpenditure;
-      this.calculatedExpenditure.foodExpenditure = this.getFoodExpenses(userParameters, budget);
-      budget = budget - this.calculatedExpenditure.foodExpenditure;
-      this.radius = this.calculateRadius(userParameters, budget);
-      navigator.geolocation.getCurrentPosition((position) => {
-        this.getApplicableLocations(this.radius, position, userParameters.duration, byRoad);
-      }, (error) => {
-        console.log('No Location Available :: ' + error);
-        this.getApplicableLocations(this.radius, defaultPosition, userParameters.duration, byRoad);
-      });
-    } catch (e) {
-      this.snackBar.open(e, '', {duration: 5000});
-    }
-  }
-
-  /**
-   * @param params
-   */
-  budgetValidations(budget) {
-    if (budget && budget > 0) {
-      return true;
-    } else {
-      throw new Error('Your budget is too low. Please add more or change parameters of search');
-    }
-  }
-
-  /**
-   *
-   * @param params
-   * @param remainingBudget
-   * @returns {any}
-   */
-  getHotelExpenses(params:UserParameters , remainingBudget) {
-    let hotelBudget;
-    hotelBudget = (Math.ceil(params.person / 2)) * constant.searchConstants.hotelAndFoodPrices[params.hotel.starrating].hotelPrice * (params.duration-1);
-    if (this.budgetValidations(remainingBudget - hotelBudget)) {
-      return hotelBudget;
-    } else {
-      /**
-       * TODO :: Show message to user
-       */
-      return false;
-    }
-  }
-
-  /**
-   *
-   * @param params
-   * @param remainingBudget
-   * @returns {any}
-   */
-  getFoodExpenses(params: UserParameters, remainingBudget) {
-    let foodBudget;
-    foodBudget = params.person * constant.searchConstants.hotelAndFoodPrices[params.hotel.starrating].foodPrice * (params.duration);
-    if (this.budgetValidations(remainingBudget - foodBudget)) {
-      return foodBudget;
-    } else {
-      /**
-       * TODO :: Show message to user
-       */
-      return false;
-    }
-
-  }
-
-  /**
-   *
-   * @param params
-   * @param remainingBudget
-   * @returns {any}
-   */
-  calculateRadius(params:UserParameters , remainingBudget) {
-    let radius;
-    let numberOfVehicles;
-    const travelConst = constant.searchConstants.modeOfTravelPrices;
-
-    this.calculatedExpenditure.travelExpenditure = remainingBudget;
-
-    switch (params.travel.travelmode) {
-
-      case constant.travelMode.fourwheeler: {
-        numberOfVehicles = Math.ceil(params.person / travelConst.driving.seatingCapacity[params.travel.cartype]);
-        radius = Math.ceil(remainingBudget / (numberOfVehicles * travelConst.driving.engineType[params.travel.enginetype.toLowerCase()]));
-        break;
-      }
-      case constant.travelMode.twowheeler: {
-        numberOfVehicles = Math.ceil(params.person / travelConst.twoWheeler.seatingCapacity);
-        radius = Math.ceil(remainingBudget / (numberOfVehicles * travelConst.twoWheeler.petrol));
-        break;
-      }
-      case constant.travelMode.bus: {
-        radius = remainingBudget / (params.person * travelConst.bus[params.travel.bustype ? params.travel.bustype.toLowerCase() : 'nonAc']);
-        break;
-      }
-      case constant.travelMode.train: {
-        radius = remainingBudget / (params.person * travelConst.train[params.travel.trainclass ? params.travel.trainclass : 1]);
-        break;
-      }
-      default: {
-        numberOfVehicles = Math.ceil(params.person / travelConst.twoWheeler.seatingCapacity);
-        radius = Math.ceil(remainingBudget / (numberOfVehicles * travelConst.twoWheeler.petrol));
-        break;
-      }
-    }
-    return (radius / 2);
-  }
-
-  /**
-   *
-   * @param radius
-   * @param position
-   * @param totalDays
-   * @param byRoad
-   */
-  getApplicableLocations(radius, position, totalDays, byRoad) {
-    
-      this.searchDataService.getApplicableLocations(radius, 
-        position, totalDays, byRoad,this.calculatedExpenditure);
-      this.availableLocationsSubs = this.searchDataService.getApplicableLocationsSubs().subscribe(data => {
-        if(data){
-          this.applicableLocations =data;
-          this.mapInitializer(this.applicableLocations, position);
-          this.applicableDestinations.next(this.applicableLocations);
-          console.log('destinations set in parent');
-        }
-      });
-  }
-
 
   /**
    *
